Move convert_pop to convert_operators.py and polish convert_pop
